+++ src/test/kotlin/domaintest/LottoSellerTest.kt
@@ -1,3 +1,5 @@
+package domaintest
+
 import domain.LottoSeller
 import org.assertj.core.api.Java6Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
@@ -6,8 +8,8 @@ import org.junit.jupiter.params.provider.ValueSource
 class LottoSellerTest {
 
     @ParameterizedTest
-    @ValueSource(ints = [1,2,3,4,5,6])
-    fun `자동 로또를 여러개 판매하기`(autoMaticLottosCount:Int){
+    @ValueSource(ints = [1, 2, 3, 4, 5, 6])
+    fun `자동 로또를 여러개 판매하기`(autoMaticLottosCount: Int) {
         assertThat(
             LottoSeller().sellAutoMaticLottos(autoMaticLottosCount).size
         ).isEqualTo(autoMaticLottosCount)
