+++ src/test/kotlin/domaintest/lottogeneratortest/AutomaticLottoGeneratorTest.kt
@@ -1,4 +1,4 @@
-package lottogenerator
+package domaintest.lottogeneratortest
 
 import domain.lottogenerator.AutomaticLottoGenerator
 import domain.model.lotto.Lotto
@@ -11,12 +11,12 @@ class AutomaticLottoGeneratorTest {
     lateinit var automaticLottoGenerator: AutomaticLottoGenerator
 
     @BeforeEach
-    fun setUp(){
+    fun setUp() {
         automaticLottoGenerator = AutomaticLottoGenerator()
     }
 
     @Test
-    fun `자동 로또를 생성하기`(){
+    fun `자동 로또를 생성하기`() {
         val lotto = automaticLottoGenerator.generate()
 
         assertThat(lotto).isInstanceOf(Lotto::class.java)
