+++ src/test/kotlin/domaintest/lottogeneratortest/WinningLottoGeneratorTest.kt
@@ -1,4 +1,4 @@
-package lottogenerator
+package domaintest.lottogeneratortest
 
 import domain.lottogenerator.WinningLottoGenerator
 import domain.model.WinningLotto
@@ -11,13 +11,13 @@ class WinningLottoGeneratorTest {
     lateinit var winningLottoGenerator: WinningLottoGenerator
 
     @BeforeEach
-    fun setUp(){
+    fun setUp() {
         winningLottoGenerator = WinningLottoGenerator()
     }
 
     @Test
-    fun `1등 로또를 생성하기`(){
-        val winningLotto = winningLottoGenerator.generateWinningLotto(listOf(1,2,3,4,5,6), 7)
+    fun `1등 로또를 생성하기`() {
+        val winningLotto = winningLottoGenerator.generateWinningLotto(listOf(1, 2, 3, 4, 5, 6), 7)
 
         assertThat(winningLotto).isInstanceOf(WinningLotto::class.java)
     }
