+++ src/test/kotlin/domaintest/modeltest/PurchasedLottosTest.kt
@@ -1,3 +1,5 @@
+package domaintest.modeltest
+
 class PurchasedLottosTest {
