+++ src/test/kotlin/utiltest/NumericValidatorTest.kt
@@ -1,7 +1,9 @@
-import util.NumericValidator
+package utiltest
+
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import util.NumericValidator
 
 class NumericValidatorTest {
